+++ src/definitions/RoleAdmin.ts
@@ -1,1 +1,7 @@
-export const ROLE_ADMIN = ["SUPER_ADMIN", "ADMIN", "SALE", "ACCOUNT"];
+export const ROLE_ADMIN = [
+  { key: "", status: "ALL" },
+  { key: "SUPER_ADMIN", status: "SUPER_ADMIN" },
+  { key: "ADMIN", status: "ADMIN" },
+  { key: "SALE", status: "SALE" },
+  { key: "ACCOUNT", status: "ACCOUNT" },
+];