+++ src/pages/admin/IndexAdmin.tsx
@@ -1,31 +1,46 @@
 import { EditOutlined } from "@ant-design/icons";
-import { Badge, Button, Pagination, Table } from "antd";
+import { Badge, Button, Pagination, Select, Table } from "antd";
+import "antd/dist/antd.css";
 import React, { useEffect, useState } from "react";
+import { string } from "yup";
 import ActionButton from "../../components/button/ActionButton";
 import AddButtton from "../../components/button/AddButton";
 import { CardContainer } from "../../components/card/CardContainer";
 import Layout from "../../components/layout/Layout";
 import { AdminDatasource } from "../../datasource/AdminDatasource";
+import { ROLE_ADMIN } from "../../definitions/RoleAdmin";
 import { STATUS_NORMAL_MAPPING } from "../../definitions/Status";
 import {
-  UserStaffEntity,
-  UserStaffEntity_INIT,
+  UserStaffPageEntity,
+  UserStaffPageEntity_INIT,
 } from "../../entities/UserStaffEntities";
 import { color } from "../../resource";
 import { DateTimeUtil } from "../../utilities/DateTimeUtil";
 
 const IndexAdmin = () => {
-  const [data, setData] = useState<UserStaffEntity[]>([UserStaffEntity_INIT]);
+  const row = 10;
+  const [data, setData] = useState<UserStaffPageEntity>(
+    UserStaffPageEntity_INIT
+  );
+  const [current, setCurrent] = useState(1);
+  const [searchStatus, setSearchStatus] = useState<boolean>();
+  const [searchRole, setSearchRole] = useState<string>();
+  const [roleNull, setRoleNull] = useState<string>();
 
   const fecthAdmin = async () => {
-    await AdminDatasource.getAdminList().then((res: UserStaffEntity[]) => {
+    await AdminDatasource.getAdminList(
+      current,
+      row,
+      searchStatus,
+      searchRole
+    ).then((res: UserStaffPageEntity) => {
       setData(res);
     });
   };
 
   useEffect(() => {
     fecthAdmin();
-  }, []);
+  }, [current, searchStatus, searchRole]);
 
   const colorStatus = (status: string) => {
     var mapStatus = STATUS_NORMAL_MAPPING[status];
@@ -34,18 +49,37 @@ const IndexAdmin = () => {
     return colorText;
   };
 
+  const onChangePage = (page: number) => {
+    setCurrent(page);
+  };
+
+  const handleOnChangeRole = (value: any) => {
+    if (value != "") {
+      setSearchRole(value);
+    } else {
+      setSearchRole(roleNull);
+    }
+  };
+
+  const handleOnChangeStatus = (value: any) => {
+    setSearchStatus(value);
+  };
+
   const columns = [
     {
       title: "ชื่อ-นามสกุล",
       dataIndex: "firstname",
       key: "firstname",
       width: "10%",
-    },
-    {
-      title: "",
-      dataIndex: "lastname",
-      key: "lastname",
-      width: "15%",
+      render: (value: any, row: any, index: number) => {
+        return {
+          children: (
+            <span>
+              {row.firstname + " " + row.lastname}
+            </span>
+          ),
+        };
+      },
     },
     {
       title: "ชื่อผู้ใช้ (User Name)",
@@ -104,7 +138,9 @@ const IndexAdmin = () => {
               <ActionButton
                 icon={<EditOutlined />}
                 color={color.primary1}
-                onClick={() => (window.location.href = "/EditAdmin/id=" + value)}
+                onClick={() =>
+                  (window.location.href = "/EditAdmin/id=" + value)
+                }
               />
             </div>
           ),
@@ -114,16 +150,43 @@ const IndexAdmin = () => {
   ];
 
   const pageTitle = (
-    <div className="container" style={{ padding: "10px" }}>
-      <div className="d-flex justify-content-between">
-        <div className="text-left">
-          <span
-            className="card-label font-weight-bolder text-dark"
-            style={{ fontSize: 22, fontWeight: "bold", padding: "8px" }}
-          >
-            <strong>รายชื่อผู้ดูแลระบบ (User Management)</strong>
-          </span>
-        </div>
+    <div
+      className="container d-flex justify-content-between"
+      style={{ padding: "10px" }}
+    >
+      <div className="col-lg-5">
+        <span
+          className="card-label font-weight-bolder text-dark"
+          style={{ fontSize: 22, fontWeight: "bold", padding: "8px" }}
+        >
+          <strong>รายชื่อผู้ดูแลระบบ (User Management)</strong>
+        </span>
+      </div>
+      <div className="col-lg-2">
+        <Select
+          className="col-lg-12"
+          defaultValue="ALL"
+          onChange={handleOnChangeRole}
+        >
+          {ROLE_ADMIN.map((item) => (
+            <option value={item.key}>{item.status}</option>
+          ))}
+        </Select>
+      </div>
+      <div className="col-lg-2">
+        <Select
+          className="col-lg-12"
+          onChange={handleOnChangeStatus}
+          defaultValue="สถานะทั้งหมด"
+        >
+          <option value="" selected={true}>
+            สถานะทั้งหมด
+          </option>
+          <option value="true">ใช้งาน</option>
+          <option value="false">ไม่ได้ใช้งาน</option>
+        </Select>
+      </div>
+      <div className="col-lg-2">
         <AddButtton
           text="เพิ่มผู้ดูแลระบบ"
           onClick={() => (window.location.href = "/AddAdmin")}
@@ -137,7 +200,7 @@ const IndexAdmin = () => {
       {pageTitle}
       <CardContainer>
         <Table
-          dataSource={data}
+          dataSource={data.results}
           columns={columns}
           pagination={false}
           size="large"
@@ -145,8 +208,13 @@ const IndexAdmin = () => {
         />
       </CardContainer>
       <div className="d-flex justify-content-between pt-5">
-        <p>รายการทั้งหมด {data.length} รายการ</p>
-        <Pagination defaultCurrent={1} total={1} />
+        <h5>รายการทั้งหมด {data.total} รายการ</h5>
+        <Pagination
+          current={current}
+          total={data.total}
+          onChange={onChangePage}
+          pageSize={row}
+        />
       </div>
     </Layout>
   );