+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Version - 0]
+## Feature [v0.1.0] 2022-07-05
+- Admin Page
+- Connect API CRUD Admin menu
 ## Feature [v0.0.1] 2022-07-04
 - Auth Page
 - Add & Edit Farmer Menu